+++ src/ComputeSharp.Graphics/Resources/Debug/TransferBufferDebugView{T}.cs
@@ -1,26 +1,26 @@
 ﻿using System;
 using System.Diagnostics;
 
-namespace ComputeSharp.Resources.Views
+namespace ComputeSharp.Resources.Debug
 {
     /// <summary>
-    /// A debug proxy used to display items in a <see cref="Buffer{T}"/> instance.
+    /// A debug proxy used to display items in a <see cref="TransferBuffer{T}"/> instance.
     /// </summary>
     /// <typeparam name="T">The type of items to display.</typeparam>
-    internal sealed class BufferDebugView<T>
+    internal sealed class TransferBufferDebugView<T>
         where T : unmanaged
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="BufferDebugView{T}"/> class with the specified parameters.
+        /// Initializes a new instance of the <see cref="TransferBufferDebugView{T}"/> class with the specified parameters.
         /// </summary>
-        /// <param name="buffer">The input <see cref="Buffer{T}"/> instance with the items to display.</param>
-        public BufferDebugView(Buffer<T>? buffer)
+        /// <param name="buffer">The input <see cref="TransferBuffer{T}"/> instance with the items to display.</param>
+        public TransferBufferDebugView(TransferBuffer<T>? buffer)
         {
             if (buffer is not null)
             {
                 var items = GC.AllocateUninitializedArray<T>(buffer.Length);
 
-                buffer.CopyTo(ref items[0], buffer.Length, 0);
+                buffer.Span.CopyTo(items);
 
                 Items = items;
             }
