+++ src/ComputeSharp.Graphics/Resources/ReadOnlyTexture3D{T,TPixel}.cs
@@ -3,7 +3,7 @@ using ComputeSharp.__Internals;
 using ComputeSharp.Exceptions;
 using ComputeSharp.Graphics.Resources.Enums;
 using ComputeSharp.Resources;
-using ComputeSharp.Resources.Views;
+using ComputeSharp.Resources.Debug;
 using static TerraFX.Interop.D3D12_FORMAT_SUPPORT1;
 
 #pragma warning disable CS0618
