+++ src/ComputeSharp.Graphics/Resources/UploadTexture2D{T}.cs
@@ -1,7 +1,7 @@
 ﻿using System.Diagnostics;
 using ComputeSharp.Graphics.Resources.Enums;
 using ComputeSharp.Resources;
-using ComputeSharp.Resources.Views;
+using ComputeSharp.Resources.Debug;
 
 namespace ComputeSharp
 {
